# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "쿠팡이츠 클론코딩",
        "contact": {
            "email": "wngur6076@naver.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "테스트 서버"
        },
        {
            "url": "https://delivery-service-web.herokuapp.com/api",
            "description": "실 서버"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "로그인 하기",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "로그인 성공"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Authorization"
                ],
                "summary": "유저정보 가져오기",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::show",
                "responses": {
                    "200": {
                        "description": "응답 성공"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "delete": {
                "tags": [
                    "Authorization"
                ],
                "summary": "로그아웃 하기",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::destroy",
                "responses": {
                    "200": {
                        "description": "로그아웃 성공"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "토큰 재발급 하기",
                "operationId": "App\\Http\\Controllers\\Auth\\RefreshController::store",
                "responses": {
                    "200": {
                        "description": "토큰 재발급 성공"
                    },
                    "401": {
                        "description": "refresh_token_error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "회원가입 하기",
                "operationId": "App\\Http\\Controllers\\Auth\\RegisterController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "회원가입 성공",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "422": {
                        "description": "validate_error"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/eateries/{id}": {
            "get": {
                "tags": [
                    "음식점"
                ],
                "summary": "특정 음식점 가져오기",
                "description": "특정 음식점 아이템을 가져온다.",
                "operationId": "getEateryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "eatery_id(1~50 test-case 있음)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/menus/{id}": {
            "get": {
                "tags": [
                    "메뉴"
                ],
                "summary": "특정 메뉴 가져오기",
                "description": "특정 메뉴 아이템을 가져온다.",
                "operationId": "getMenuById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "menu_id(메뉴아이디 적어주세요.)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}